+++ src/aerospike_clj/utils.clj
@@ -60,3 +60,13 @@
   "An optimized way to convert vectors into Java arrays of type `cls`"
   [cls v]
   (.toArray ^Collection v ^"[Ljava.lang.Object;" (make-array cls (count v))))
+
+(defn vectorize 
+  "convert a single value to a vector or any collection to the equivalent vector.
+  NOTE: a map or a set have no defined order so vectorize them is not allowed"
+  [v]
+  (cond
+    (or (map? v) (set? v)) (throw (IllegalArgumentException. "undefined sequence order for argument"))
+    (or (nil? v) (vector? v) (seq? v)) (vec v)
+    :else [v]))
+