+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#### VERSION 0.6.0
+* Support ClientEvents vector to be a vector of completions instead of a single one.
+* Bump aerospike lib to 4.4.15
+
 #### VERSION 0.5.1
 * Added batch-exsits
 * Bump aerospike lib to 4.4.10